Got all of the dependencies compiling without errors.
NOTE: When running Configure, it is important to make sure snapshot is
not enabled, as it will fail to compile. Do this with the command:
./configure --without-snapshot

TODO: Move the new OpenSSL flags "-DOPENSSL_NO_GOST" and "-DOPENSSL_NO_HW_PADLOCK
(which are currently in tools/gp/pylib/gyp/generator/make.py) to the proper
OpenSSL gyp files.